% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascaris_model.R
\name{ascaris_model}
\alias{ascaris_model}
\title{Ascaris model}
\usage{
ascaris_model(M, k)
}
\arguments{
\item{M}{Number of worms}

\item{k}{Work clumping coeficient}
}
\value{
Prevalence
}
\description{
Mechanistic model of Ascaris prevalence.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
