% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ESS.R
\name{calculate_ESS}
\alias{calculate_ESS}
\title{Calculates Effective Sample Sizes and weights}
\usage{
calculate_ESS(prev, sim, w1, delta)
}
\arguments{
\item{prev}{Matrix with observed prevalences}

\item{sim}{Vector of simulated prevalences}

\item{w1}{Weight matrix 1}

\item{delta}{Threshold value to compare similarity between prevalences}
}
\value{
Vector with ESS and weight mnatrix
}
\description{
AMIS.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
