% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-data.R
\docType{data}
\name{coord}
\alias{coord}
\title{Coordinates and mean prevalence of Ascaris in Ethiopia}
\format{
An object of class \code{data.frame} with 54160 rows and 4 columns.
}
\usage{
data(coord)
}
\description{
Coordinates and mean prevalence of Ascaris in Ethiopia
}
\examples{
data(coord)
}
\keyword{datasets}
