% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprop0_ascaris.R
\name{dprop0_ascaris}
\alias{dprop0_ascaris}
\title{Prior density for Ascaris}
\usage{
dprop0_ascaris(a, b, fit.v, fit.hess, fit.inflation.factor)
}
\arguments{
\item{a}{Proposed value of M}

\item{b}{Proposed value of k}

\item{fit.v}{Fitted relationship between M and k}

\item{fit.hess}{Fitted Hesian matrix}
}
\value{
Probability
}
\description{
AMIS.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
