% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMclust.R
\name{fitMclust}
\alias{fitMclust}
\title{Fit clusters of mixtures}
\usage{
fitMclust(xx, modelName = "VVV", G = G)
}
\arguments{
\item{xx}{Sampled parameters}

\item{G}{Sequence of clusters}
}
\value{
Parameters of mixture
}
\description{
AMIS.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
