% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.R
\name{H}
\alias{H}
\title{H}
\usage{
H(x)
}
\arguments{
\item{x}{Vector of numbers}
}
\value{
A vector of 0 and 1s
}
\description{
Transform logical input to a vector of numbers.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
