% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIV_model.R
\name{HIV_model}
\alias{HIV_model}
\title{Plot Ascaris AMIS results}
\usage{
HIV_model(N0, f0, t0, phi, mu, lambda0, dt, TT)
}
\arguments{
\item{N0}{Initial population size}

\item{f0}{Fraction of initially infeceted}

\item{t0}{Year HIV outreak started}
}
\value{
Prevalence
}
\description{
Plot Ascaris AMIS results
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
