% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Malaria_model.R
\name{Malaria_model}
\alias{Malaria_model}
\title{Runs OpenMalaria model}
\usage{
Malaria_model(ibra, sigma2i)
}
\arguments{
\item{ibra}{Infectious bites rate per year}

\item{sigma2i}{Host inter-variability}
}
\value{
Prevalence
}
\description{
Runs OpenMalaria model
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
