% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclustMix.R
\name{mclustMix}
\alias{mclustMix}
\title{Mixture distribution}
\usage{
mclustMix(G = 1:10)
}
\arguments{
\item{G}{Sequence of clusters}
}
\value{
Parameters of mixture
}
\description{
AMIS.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
