% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtComp.R
\name{mvtComp}
\alias{mvtComp}
\title{Proposal for mixtur distribution}
\usage{
mvtComp(df = 3)
}
\arguments{
\item{df}{Number of degrees of freedom, df=3 for t Student distribution.}
}
\value{
Distribution
}
\description{
Mechanistic model of Ascaris prevalence.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
