% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncho_model.R
\name{Oncho_model}
\alias{Oncho_model}
\title{Runs EPIONCHO model under two alternative control scenarious}
\usage{
Oncho_model(abr, k)
}
\arguments{
\item{abr}{Annual biting rate}

\item{k}{Host variability contant}
}
\value{
Prevalence
}
\description{
Runs EPIONCHO model under two alternative control scenarious
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
