% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AMIS_Ascaris.R
\name{plot_AMIS_Ascaris}
\alias{plot_AMIS_Ascaris}
\title{Plot Ascaris AMIS results}
\usage{
plot_AMIS_Ascaris(coord, ans)
}
\arguments{
\item{coord}{Coordinates od pixels in Ethiopia}

\item{p}{Fraction of initially infeceted}

\item{t0}{Year HIV outreak started}
}
\value{
Prevalence
}
\description{
Plot Ascaris AMIS results
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
