% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev-data.R
\docType{data}
\name{prev}
\alias{prev}
\title{Ascaris prevalence distribution for 10km x 10km gridded map}
\format{
An object of class \code{matrix} with 11369 rows and 100 columns.
}
\usage{
data(prev)
}
\description{
Ascaris prevalence distribution for 10km x 10km gridded map
}
\examples{
data(prev)
}
\keyword{datasets}
