% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_AMIS_ascaris.R
\name{run_AMIS_ascaris}
\alias{run_AMIS_ascaris}
\title{Run AMIS for Ascaris}
\usage{
run_AMIS_ascaris(prev, n.param, NN, delta, ESS.R)
}
\arguments{
\item{prev}{Matrix with observed prevalences: row for a separate pixel}

\item{n.param}{Number of parameters}

\item{NN}{Vector containing number of samples for each iteration}

\item{delta}{Threshold for prevalence}

\item{ESS.R}{Minimum Effective Sample Size}
}
\value{
Matrix with sampled vectpors and weigts for each pixel
}
\description{
Parameter estimation for Adcaris prevalence distributions.
}
\examples{
run_AMIS_ascaris(prev=prev, n.param=2, NN=rep(1000, 100), delta=5, ESS.R=2000)

}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
